Remove some sections in problem analysis
I'm proposing the following changes:

Do we need to go in depth here? This can be dealt with in sprint 2 or 3.

> Persistent REPL environment
> Plug-in development in Eclipse

If we go for whole program execution first, this isn't needed either:

> Showing the contents of the environment

If we don't well first have to figure out how directly invoking the
HybridInterpreter works *and* be able to parse partial programs before
we know how to tackle this problem.

Might be worthwhile to pursue IPython here, but can also be dealt with
in sprint 2 or 3.

> Extending the product with literate programming