Initial structure of final report with copied sections
The directory structure is such that each chapter has a .tex file in the
"root" project directory, and for each chapter there's also a
subdirectory containing all the sections of that chapter.

E.g. The first chapter, "Background", has a file background.tex and a
subdirectory background/ with the Spoofax section and the REPL section.